+++ quint/src/cli.ts
@@ -150,8 +150,8 @@ const compileCmd = {
 
 // construct repl commands with yargs
 const replCmd = {
-  command: ['repl', '*'],
-  desc: 'run an interactive Read-Evaluate-Print-Loop',
+  command: ['repl [commands..]', '*'],
+  desc: 'Run an interactive Read-Evaluate-Print-Loop. Optionally, takes one or more commands to execute upon entering the REPL.',
   builder: (yargs: any) =>
     yargs
       .option('require', {
@@ -403,6 +403,12 @@ const docsCmd = {
 const validate = (argv: any, opts: any) => {
   // Validate that only array options can be specified more than once.
   for (const key in argv) {
+    if (key == 'commands' && (argv['_'][0] || 'repl') === 'repl') {
+      // Skip checking repl's positional arguments for both `quint ...` and
+      // `quint repl ...` commands. Note that `opts` don't have enought information
+      // on positional arguments, thus making this special case necessary.
+      continue
+    }
     if (key !== '_' && !opts.array.includes(key) && Array.isArray(argv[key])) {
       throw new Error(`--${key} can not be specified more than once`)
     }